feat: migrate from streaming_wrapper to async_client
